Move Kuramoto NN to seperate file
