+++ models/Kuramoto/__init__.py
@@ -1,4 +1,5 @@
 from .ABM import Kuramoto_ABM
 from .DataGeneration import get_data
 from .Langevin import Kuramoto_Langevin_sampler
+from .NN import Kuramoto_NN
 from .Regression import rank, regression
